Add Vary, consider intranet as connected
