+++ geoportal/c2cgeoportal_geoportal/lib/caching.py
@@ -28,12 +28,9 @@
 
 import inspect
 import logging
-from enum import Enum
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, cast
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional
 
 import pyramid.interfaces
-import pyramid.request
-import pyramid.response
 import sqlalchemy.ext.declarative
 import zope.interface
 from dogpile.cache.api import NO_VALUE, CacheBackend
@@ -172,135 +169,3 @@ class HybridRedisSentinelBackend(HybridRedisBackend):
     def __init__(self, arguments: Dict[str, Any]):
         super().__init__(arguments)
         self._redis = RedisSentinelBackend(arguments)
-
-
-class Cache(Enum):
-    """Enumeration for the possible cache values."""
-
-    # For responses that can be put in a public cache (proxy+browser)
-    PUBLIC = 0
-    # For responses that can be put in a public cache (proxy+browser),
-    # But with a smalls max-age (10s), for tow possible reason:
-    # - The resource can change and the browser need to get the new version
-    # - The chance of cache hit is too small to put it in a cache
-    # We use a small cache instead of no cache to protect the service against too high traffic.
-    PUBLIC_NO = 1
-    # For responses that can be put in a private cache (browser)
-    PRIVATE = 2
-    # See PUBLIC_NO and PRIVATE
-    PRIVATE_NO = 3
-
-
-CORS_METHODS = "HEAD, GET, POST, PUT, DELETE"
-
-
-def _set_cors_headers(
-    request: pyramid.request.Request,
-    response: pyramid.response.Response,
-    service_name: str,
-    service_headers_settings: Dict[str, Any],
-    credentials: bool,
-) -> None:
-    """Handle CORS requests, as specified in https://www.w3.org/TR/cors/."""
-    response.vary = (response.vary or ()) + ("Origin",)
-
-    if "Origin" not in request.headers:
-        return  # Not a CORS request if this header is missing
-    origin = request.headers["Origin"]
-
-    if request.method == "OPTIONS" and "Access-Control-Request-Method" not in request.headers:
-        LOG.warning("CORS preflight query missing the Access-Control-Request-Method header")
-        return
-
-    allowed_origins = cast(List[str], service_headers_settings.get("access_control_allow_origin", []))
-    if origin not in allowed_origins:
-        if "*" in allowed_origins:
-            origin = "*"
-            credentials = False  # Force no credentials
-        else:
-            LOG.warning("CORS query not allowed for origin=%s, service=%s", origin, service_name)
-            return
-
-    response.headers.update(
-        {"Access-Control-Allow-Origin": origin, "Access-Control-Allow-Methods": CORS_METHODS}
-    )
-
-    max_age = service_headers_settings.get("access_control_max_age", 3600)
-    response.headers["Access-Control-Max-Age"] = str(max_age)
-
-    if credentials:
-        response.headers["Access-Control-Allow-Credentials"] = "true"
-
-    if request.method != "OPTIONS":
-        return
-
-    response.cache_control.max_age = max_age
-
-    if not service_headers_settings or "access_control_allow_origin" not in service_headers_settings:
-        LOG.warning("CORS query not configured for service=%s", service_name)
-        return
-
-    requested_headers = request.headers.get("Access-Control-Request-Headers", False)
-    if requested_headers:
-        # For the moment, we allow all requested headers
-        response.headers["Access-Control-Allow-Headers"] = requested_headers
-
-    # If we start using headers in responses, we'll have to add
-    # Access-Control-Expose-Headers
-
-
-def _set_common_headers(
-    request: pyramid.request.Request,
-    response: pyramid.response.Response,
-    service_headers_settings: Dict[str, Dict[str, str]],
-    cache: Cache,
-    content_type: Optional[str],
-) -> pyramid.response.Response:
-    """Set the common headers."""
-
-    response.headers.update(service_headers_settings.get("headers", {}))
-
-    maxage = (
-        service_headers_settings.get("cache_control_max_age", 3600)
-        if cache in (Cache.PUBLIC, Cache.PRIVATE)
-        else service_headers_settings.get("cache_control_max_age_nocache", 10)
-    )
-    response.cache_control.max_age = maxage
-    if maxage == 0:
-        response.cache_control.no_cache = True
-        response.cache_control.no_store = True
-    elif cache in (Cache.PUBLIC, Cache.PUBLIC_NO):
-        response.cache_control.public = True
-    elif cache in (Cache.PRIVATE, Cache.PRIVATE_NO):
-        if hasattr(request, "user") and request.user is not None:
-            response.cache_control.private = True
-        else:
-            response.cache_control.public = True
-    else:
-        raise Exception("Invalid cache type")
-
-    if content_type is not None:
-        response.content_type = content_type
-
-    return response
-
-
-def set_common_headers(
-    request: pyramid.request.Request,
-    service_name: str,
-    cache: Cache,
-    response: pyramid.response.Response = None,
-    credentials: bool = True,
-    content_type: Optional[str] = None,
-) -> pyramid.response.Response:
-    """Set the common headers."""
-    if response is None:
-        response = request.response
-
-    headers_settings = request.registry.settings.get("headers", {})
-    service_headers_settings = headers_settings.get(service_name, {})
-
-    _set_cors_headers(request, response, service_name, service_headers_settings, credentials)
-    if request.method == "OPTIONS":
-        return response
-    return _set_common_headers(request, response, service_headers_settings, cache, content_type)
