+++ geoportal/c2cgeoportal_geoportal/views/dynamic.py
@@ -38,7 +38,8 @@ from sqlalchemy import func
 from c2cgeoportal_commons import models
 from c2cgeoportal_commons.models import main
 from c2cgeoportal_geoportal.lib.cacheversion import get_cache_version
-from c2cgeoportal_geoportal.lib.caching import Cache, get_region, set_common_headers
+from c2cgeoportal_geoportal.lib.caching import get_region
+from c2cgeoportal_geoportal.lib.common_headers import Cache, set_common_headers
 
 CACHE_REGION = get_region("std")
