+++ geoportal/c2cgeoportal_geoportal/views/layers.py
@@ -60,7 +60,8 @@ from sqlalchemy.sql import and_, or_
 
 from c2cgeoportal_commons import models
 from c2cgeoportal_geoportal.lib import get_roles_id
-from c2cgeoportal_geoportal.lib.caching import Cache, get_region, set_common_headers
+from c2cgeoportal_geoportal.lib.caching import get_region
+from c2cgeoportal_geoportal.lib.common_headers import Cache, set_common_headers
 from c2cgeoportal_geoportal.lib.dbreflection import _AssociationProxy, get_class, get_table
 
 if TYPE_CHECKING:
