+++ geoportal/c2cgeoportal_geoportal/views/login.py
@@ -53,7 +53,8 @@ from c2cgeoportal_commons.lib.email_ import send_email_config
 from c2cgeoportal_commons.models import static
 from c2cgeoportal_geoportal import is_valid_referrer
 from c2cgeoportal_geoportal.lib import get_setting, is_intranet, oauth2
-from c2cgeoportal_geoportal.lib.caching import Cache, get_region, set_common_headers
+from c2cgeoportal_geoportal.lib.caching import get_region
+from c2cgeoportal_geoportal.lib.common_headers import Cache, set_common_headers
 from c2cgeoportal_geoportal.lib.functionality import get_functionality
 
 LOG = logging.getLogger(__name__)
