+++ geoportal/c2cgeoportal_geoportal/views/printproxy.py
@@ -39,7 +39,8 @@ from pyramid.view import view_config
 
 from c2cgeoportal_commons.lib.url import Url
 from c2cgeoportal_geoportal.lib import is_intranet
-from c2cgeoportal_geoportal.lib.caching import Cache, get_region
+from c2cgeoportal_geoportal.lib.caching import get_region
+from c2cgeoportal_geoportal.lib.common_headers import Cache
 from c2cgeoportal_geoportal.lib.functionality import get_functionality
 from c2cgeoportal_geoportal.views.proxy import Proxy
