+++ geoportal/c2cgeoportal_geoportal/views/vector_tiles.py
@@ -35,7 +35,7 @@ from tilecloud import TileCoord
 from tilecloud.grid.free import FreeTileGrid
 
 from c2cgeoportal_commons.models import DBSession, main
-from c2cgeoportal_geoportal.lib.caching import Cache, set_common_headers
+from c2cgeoportal_geoportal.lib.common_headers import Cache, set_common_headers
 
 LOG = logging.getLogger(__name__)
