+++ geoportal/tests/functional/test_oauth2.py
@@ -133,7 +133,7 @@ class TestLoginView(TestCase):
         response = Login(request).oauth2token()
         assert response.headers["Content-Type"] == "application/json"
         assert response.headers["Pragma"] == "no-cache"
-        assert response.headers["Vary"] == "Origin"
+        assert response.headers["Vary"] == "Origin, Cookie"
         assert response.headers["Cache-Control"] == "max-age=10, no-store, public"
         data = json.loads(response.body)
         assert set(data.keys()) == {"access_token", "expires_in", "token_type", "refresh_token"}
